plugins/rgb/fill_screen/requirements.txt
numpy==2.2.6
pillow==11.2.1
pluggy==1.6.0